ATR_analysis: get the correct time for cache ATR list
wget --output-document=foo will set the date (and time) of the file foo
to the date of the remote file. But we want the file to have the date
the wget was run and not the date the remote file was last updated.

We new use get $url --output-file=foo

Problem detected on Debian jessie (8.1)