ATR_analysis: correctly use wget to store the ATR list
Using --output-file= correctly set the file time. But the content of the
file is the log of the wget command not the content of the remote file.
The code was completely broken.

This change reverts a4165dbd2baca4588c4ac8b21bc40a6dc21c54d8 and add the
missing touch(1) call.

Closes Debian bug #804049: [pcsc-tools] Update of smartcard_list.txt is
broken