scriptor: fix -u option to make stdout unbuffered
" After submitting the patch I noticed that Perl's getopt does not properly
support flags like -u (which I had assumed to be supported from the current
scriptor code). According to Getopt/Std.pm:
    getopt('oDI', \%opts);    # -o, -D & -I take arg.  Values in %opts

This went unnoticed during my tests as -u was always passed as the last
argument to scriptor.

My suggestion would be to apply the patch below on top of the last commit
to have proper -u support. Please note that getopts() outputs a warning
on stderr if unknown options are given which getopt() does not. If this
is not wanted code like

$SIG{__WARN__} = sub {};
getopts ("r:p:u" , \%options);
undef $SIG{__WARN__};

would have to be used. "

Thanks to Reinhard Speyerer for the patch.