pcsc_scan: Exit if no reader is found and -c or -r is used
The command was waiting for a first reader. But because the execution
is in quiet mode when -r or -c is used the command did not displayed the
classic message:
Using reader plug'n play mechanism
Scanning present readers...
Waiting for the first reader...
and was waiting.

Now we display "No reader found." and exit.