Replace hardcoded path in ATR_analysis and ATR_analysis.1p
The path "/usr/local/share/pcsc/" is replaced by the value $pcsc_dir.
pcsc_dir is defined as $datadir/pcsc and datadir is $prefix/share.

This allows to correctly use and document the localtion of
smartcard_list.txt when $prefix is not set as the standards /usr or
/usr/local.
For example it is needed for GNU Guix which uses a non-standard prefix.

Thanks to Jake Leporte for the initial patch.